Aastu.mobile.g3.mihret.data layer

+++ mobile/AASTU-mobile-group-3/ecommerse_collab/lib/features/product/presentation/pages/add_page.dart
@@ -134,7 +134,7 @@ class _AddProductState extends State<AddProduct> {
                                 id: '',
                                 name: name.text,
                                 category: category.text,
-                                price: int.parse(price.text),
+                                price: double.parse(price.text),
                                 description: description.text,
                                 image: _image?.path ?? "",
                                 seller: widget.user,
